Harden null/boundary handling and fix int-overflow in score totals
- Team distinguishes null (NullPointerException) from blank
  (IllegalArgumentException) team names, per standard Java convention.
- InMemoryScoreBoard explicitly rejects a null MatchId with
  NullPointerException for updateScore/finishMatch; getMatch stays
  permissive (null -> Optional.empty()) since 'not found' is already its
  documented non-exceptional outcome. ScoreBoard's Javadoc spells out
  both contracts.
- Score.total()/Match.totalScore() now widen to long: two individually
  valid int scores (each up to Integer.MAX_VALUE) summed as int would
  silently overflow and corrupt getSummary()'s ordering. Comparator
  updated to comparingLong accordingly.
- Tests: null team names, null matchId across update/finish/getMatch,
  a large-score boundary case, and a regression test proving summary
  ordering survives scores that would overflow a 32-bit sum.

+++ src/main/java/com/sportradar/fwcscoreboard/model/Match.java
@@ -28,7 +28,7 @@ public record Match(MatchId id, Team homeTeam, Team awayTeam, Score score, long 
         return new Match(id, homeTeam, awayTeam, newScore, startOrder);
     }
 
-    public int totalScore() {
+    public long totalScore() {
         return score.total();
     }
 }
